asdhjv

+++ src/main/java/com/caoccao/javet/interop/IV8Native.java
@@ -268,7 +268,7 @@ public interface IV8Native {
 
     Object objectGetPrivateProperty(long v8RuntimeHandle, long v8ValueHandle, int v8ValueType, String propertyName);
 
-    long objectGetPrivatePropertyA(long v8RuntimeHandle, long v8ValueHandle, int v8ValueType, String propertyName);
+    int objectGetPrivatePropertyA(long v8RuntimeHandle, long v8ValueHandle, int v8ValueType, String propertyName);
 
     Object objectGetProperty(long v8RuntimeHandle, long v8ValueHandle, int v8ValueType, Object key);
