+++ src/main/java/com/caoccao/javet/values/reference/V8ValueObject.java
@@ -526,7 +526,7 @@ public class V8ValueObject extends V8ValueReference implements IV8ValueObject {
     }
 
     @Override
-    public long getPrivatePropertyA(String propertyName) throws JavetException {
+    public int getPrivatePropertyA(String propertyName) throws JavetException {
         return checkV8Runtime().getV8Internal().objectGetPrivatePropertyA(
                 this, Objects.requireNonNull(propertyName));
     }
